+++ changelog.md
@@ -2,6 +2,10 @@
 ## Week 07
 
 ### Monday 15/April
+- Polishing the start scene [Piumi]
+- Polishing the appearance of the player [Piumi]
+
+### Monday 15/April
 - Feedback: Good job adding comments
 - Feedback: Watch your variable names, and ensure they're understandable. E.g. I don't know 're' in "userPasswordRe".